konsi_worker/requirements.txt
fastapi
flake8
gunicorn
SQLAlchemy==1.4.41
sqlmodel
uvicorn
psycopg2-binary==2.9.3
asyncpg==0.26.0
email-validator
celery>=5.2
requests==2.28.2
python-dotenv
pytest==7.2.2
redis>=4.0
Faker==15.1.1